+++ node/src/components/consensus/protocols/highway/tests.rs
@@ -66,11 +66,11 @@ where
         .collect::<Vec<_>>();
     let chainspec = new_test_chainspec(weights.clone());
     let config = Config {
+        max_execution_delay: 3,
         highway: HighwayConfig {
             pending_vertex_timeout: "1min".parse().unwrap(),
             standstill_timeout: Some(STANDSTILL_TIMEOUT.parse().unwrap()),
             log_participation_interval: Some("10sec".parse().unwrap()),
-            max_execution_delay: 3,
             ..HighwayConfig::default()
         },
         ..Default::default()