+++ node/CHANGELOG.md
@@ -30,6 +30,8 @@ All notable changes to this project will be documented in this file.  The format
 * Add a new parameter to `info_get_deploys` JSON-RPC, `finalized_approvals` - controlling whether the approvals returned with the deploy should be the ones originally received by the node, or overridden by the approvals that were finalized along with the deploy.
 * Add a `[consensus.simple_consensus]` section to `config.toml` for the simple consensus protocol.
 * Add a `consensus_protocol` setting to the chainspec to choose a consensus protocol, and a `minimum_block_time` setting for the minimum difference between a block's timestamp and its child's.
+* Move `finality_threshold_fraction` from the `highway` to the `core` section in the chainspec.
+* Move `max_execution_delay` from the `highway` to the `consensus` section in the `config.toml`.
 
 ### Changed
 * Detection of a crash no longer triggers DB integrity checks to run on node start; the checks can be triggered manually instead.